Move model content to new authors.py module
