+++ physionet-django/project/modelcomponents/generic.py
@@ -1,8 +1,9 @@
 import ckeditor.fields
 
 import bleach
-
 from django.conf import settings
+from django.db import models
+
 from project.utility import LinkFilter
 
 
@@ -85,3 +86,18 @@ class SafeHTMLField(ckeditor.fields.RichTextField):
         value = lf.convert(value)
 
         return super().clean(value, model_instance)
+
+
+class BaseInvitation(models.Model):
+    """
+    Base class for authorship invitations and storage requests
+    """
+    project = models.ForeignKey('project.ActiveProject',
+        related_name='%(class)ss', on_delete=models.CASCADE)
+    request_datetime = models.DateTimeField(auto_now_add=True)
+    response_datetime = models.DateTimeField(null=True)
+    response = models.NullBooleanField(null=True)
+    is_active = models.BooleanField(default=True)
+
+    class Meta:
+        abstract = True
